requirements.txt
Django==5.2
Pillow==11.1.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.4 